Move DrawerNavigation files to the new 'drawer_menu' package
