+++ app/src/main/java/ecb/manifest/kowalski/main/MainActivity.kt
@@ -11,9 +11,9 @@ import androidx.compose.material.icons.outlined.DirectionsCar
 import androidx.compose.material.icons.outlined.Place
 import androidx.compose.material.icons.outlined.QrCode
 import dagger.hilt.android.AndroidEntryPoint
-import ecb.manifest.kowalski.main.ui.navigation.DrawerNavigationItem
+import ecb.manifest.kowalski.main.ui.drawer_menu.DrawerNavigationItem
 import ecb.manifest.kowalski.main.ui.theme.ManifestKowalskiTheme
-import ecb.manifest.kowalski.main.ui.navigation.DrawerNavigationMenu
+import ecb.manifest.kowalski.main.ui.drawer_menu.DrawerNavigationMenu
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
